requirements.txt
Flask
psycopg2-binary
selenium
webdriver-manager
concurrent-log-handler